requirements.txt
python-docx
nltk
matplotlib
pandas
openpyxl
